# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentio SMTP API",
    "description": "Multi-tenant SMTP platform API",
    "contact": {
      "name": "Truespar Team"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.5"
  },
  "paths": {
...
    "/v1/messages/ingest": {
      "post": {
        "tags": [
          "Messages"
        ],
        "operationId": "ingest_message",
        "requestBody": {
          "description": "JSON envelope + RFC 822 body, or Content-Type: message/rfc822 with X-Envelope-* headers",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IngestResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_IngestResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "queue_id",
              "status"
            ],
            "properties": {
              "id": {
                "$ref": "#/components/schemas/MessageId"
              },
              "queue_id": {
                "type": "string"
              },
              "status": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "IngestRequest": {
        "type": "object",
        "required": [
          "envelope_from",
          "envelope_to"
        ],
        "properties": {
          "client_ip": {
            "type": [
              "string",
              "null"
            ],
            "description": "Original SMTP peer IP for SPF. Defaults to the HTTP client address."
          },
          "envelope_from": {
            "type": "string",
            "description": "SMTP MAIL FROM. Empty string is the null sender `<>`."
          },
          "envelope_to": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "SMTP RCPT TO recipients. Every domain must be hosted by this tenant."
          },
          "helo_domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "EHLO/HELO name from the original SMTP session, if known."
          },
          "raw": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 822 message as a JSON string."
          },
          "raw_base64": {
            "type": [
              "string",
              "null"
            ],
            "description": "Base64-encoded RFC 822 when `raw` is not valid UTF-8."
          }
        }
      },
...
      "MessageId": {
        "type": "string",
        "format": "uuid"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}